[migrate_parsetree] Update to 407 ast and workaround metaquot
`ocaml-migrate-parsetree` has deprecated unqualified access, however
the metaquot ppx generates unqualified references; thus we open the
modules in a different fashion as to avoid this problem and have a
clean dev build.

We also migrate to the latest AST version that we can support without
migration of the `Types` module.

This is the continuation of #38.