Generate a compat module for 4.08 and 4.09
For Types, values are shallowly migrated to 4.07 versions of
`Types.module_type` and `Types.signature_item`.
The code to convert for `Types` to 4.07 versions is selected by dune
according to the compiler version.

In 4.09, the argument passed to init_path is changed, so the right piece
of code is selected.